requirements - Copy.txt
streamlit
ultralytics
Pillow
numpy
opencv-python-headless  